Reworked entire system to have one main file
